NXP-25633: Bulk refactor without changing logic
Use a message for document bucket ids.
Externalize Computations from Processors.
Misc cleaning.
