+++ nuxeo-core/nuxeo-core-bulk/src/main/java/org/nuxeo/ecm/core/bulk/BulkCodecs.java
@@ -19,6 +19,10 @@
 
 package org.nuxeo.ecm.core.bulk;
 
+import org.nuxeo.ecm.core.bulk.message.BulkBucket;
+import org.nuxeo.ecm.core.bulk.message.BulkCommand;
+import org.nuxeo.ecm.core.bulk.message.BulkCounter;
+import org.nuxeo.ecm.core.bulk.message.BulkStatus;
 import org.nuxeo.lib.stream.codec.Codec;
 import org.nuxeo.runtime.api.Framework;
 import org.nuxeo.runtime.codec.CodecService;
@@ -36,15 +40,20 @@ public class BulkCodecs {
         // utility class
     }
 
-    public static Codec<BulkCommand> getBulkCommandCodec() {
+    public static Codec<BulkCommand> getCommandCodec() {
         return Framework.getService(CodecService.class).getCodec(DEFAULT_CODEC, BulkCommand.class);
     }
 
-    public static Codec<BulkStatus> getBulkStatusCodec() {
+    public static Codec<BulkStatus> getStatusCodec() {
         return Framework.getService(CodecService.class).getCodec(DEFAULT_CODEC, BulkStatus.class);
     }
 
-    public static Codec<BulkCounter> getBulkCounterCodec() {
+    public static Codec<BulkCounter> getCounterCodec() {
         return Framework.getService(CodecService.class).getCodec(DEFAULT_CODEC, BulkCounter.class);
     }
+
+    public static Codec<BulkBucket> getBucketCodec() {
+        return Framework.getService(CodecService.class).getCodec(DEFAULT_CODEC, BulkBucket.class);
+    }
+
 }
