+++ nuxeo-core/nuxeo-core-bulk/src/main/java/org/nuxeo/ecm/core/bulk/message/BulkStatus.java
@@ -17,7 +17,7 @@
  *     Funsho David
  */
 
-package org.nuxeo.ecm.core.bulk;
+package org.nuxeo.ecm.core.bulk.message;
 
 import java.io.Serializable;
 import java.time.Instant;
@@ -27,10 +27,11 @@ import java.util.Map;
 
 import org.apache.avro.reflect.AvroEncode;
 import org.nuxeo.ecm.core.bulk.io.InstantAsLongEncoding;
-import org.nuxeo.ecm.core.bulk.io.MapAsJsonAsStringEncoding;
+import org.nuxeo.ecm.core.bulk.message.MapAsJsonAsStringEncoding;
+import org.nuxeo.ecm.core.bulk.message.BulkCommand;
 
 /**
- * This object holds the current bulk command execution status.
+ * A message representing a command status.
  * <p/>
  * This aggregates status and metrics of documentSet creation and action computation.
  *
@@ -38,7 +39,7 @@ import org.nuxeo.ecm.core.bulk.io.MapAsJsonAsStringEncoding;
  */
 public class BulkStatus implements Serializable {
 
-    private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 20181021L;
 
     /**
      * Possible states of bulk execution.
@@ -54,7 +55,7 @@ public class BulkStatus implements Serializable {
         RUNNING,
 
         /** System has finished to scroll. */
-        COMPLETED;
+        COMPLETED
 
     }
 
@@ -79,7 +80,7 @@ public class BulkStatus implements Serializable {
     protected Map<String, Serializable> result = new HashMap<>();
 
     /**
-     * Gets bulk id.
+     * Gets command id.
      *
      * @return the id
      */
