+++ nuxeo-core/nuxeo-core-bulk/src/main/resources/OSGI-INF/bulk-config.xml
@@ -11,7 +11,7 @@
 
   <extension target="org.nuxeo.runtime.stream.service" point="streamProcessor">
     <!-- Processor used by BulkService to scroll DB -->
-    <streamProcessor name="bulkDocumentScroller" class="org.nuxeo.ecm.core.bulk.StreamBulkProcessor"
+    <streamProcessor name="bulkDocumentScroller" class="org.nuxeo.ecm.core.bulk.BulkProcessor"
                      logConfig="bulk" defaultConcurrency="1" defaultPartitions="1">
       <option name="scrollBatchSize">100</option>
       <option name="scrollKeepAlive">60</option>
