+++ nuxeo-core/nuxeo-core-bulk/src/test/java/org/nuxeo/ecm/core/bulk/io/TestBulkCommandJson.java
@@ -23,7 +23,7 @@ import java.io.File;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.nuxeo.common.utils.FileUtils;
-import org.nuxeo.ecm.core.bulk.BulkCommand;
+import org.nuxeo.ecm.core.bulk.message.BulkCommand;
 import org.nuxeo.ecm.core.bulk.CoreBulkFeature;
 import org.nuxeo.ecm.core.io.marshallers.json.AbstractJsonBijectionTest;
 import org.nuxeo.runtime.test.runner.Features;
