+++ nuxeo-core/nuxeo-core-bulk/src/test/java/org/nuxeo/ecm/core/bulk/io/TestBulkCommandJsonReader.java
@@ -28,7 +28,7 @@ import java.util.Map;
 
 import org.junit.Test;
 import org.nuxeo.common.utils.FileUtils;
-import org.nuxeo.ecm.core.bulk.BulkCommand;
+import org.nuxeo.ecm.core.bulk.message.BulkCommand;
 import org.nuxeo.ecm.core.bulk.CoreBulkFeature;
 import org.nuxeo.ecm.core.io.marshallers.json.AbstractJsonReaderTest;
 import org.nuxeo.runtime.test.runner.Features;
