+++ nuxeo-core/nuxeo-core-bulk/src/test/java/org/nuxeo/ecm/core/bulk/io/TestBulkCommandJsonWriter.java
@@ -30,7 +30,7 @@ import static org.nuxeo.ecm.core.io.registry.MarshallingConstants.ENTITY_FIELD_N
 import java.util.HashMap;
 
 import org.junit.Test;
-import org.nuxeo.ecm.core.bulk.BulkCommand;
+import org.nuxeo.ecm.core.bulk.message.BulkCommand;
 import org.nuxeo.ecm.core.bulk.CoreBulkFeature;
 import org.nuxeo.ecm.core.io.marshallers.json.AbstractJsonWriterTest;
 import org.nuxeo.ecm.core.io.marshallers.json.JsonAssert;
