+++ nuxeo-core/nuxeo-core-bulk/src/test/java/org/nuxeo/ecm/core/bulk/io/TestMapAsJsonAsStringEncoding.java
@@ -32,6 +32,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.nuxeo.ecm.core.bulk.CoreBulkFeature;
+import org.nuxeo.ecm.core.bulk.message.MapAsJsonAsStringEncoding;
 import org.nuxeo.runtime.test.runner.Features;
 import org.nuxeo.runtime.test.runner.FeaturesRunner;
