+++ nuxeo-features/nuxeo-search/nuxeo-search-rest-api/src/test/java/org/nuxeo/ecm/restapi/server/jaxrs/search/test/BulkActionTest.java
@@ -22,7 +22,7 @@ package org.nuxeo.ecm.restapi.server.jaxrs.search.test;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
-import static org.nuxeo.ecm.core.bulk.BulkStatus.State.COMPLETED;
+import static org.nuxeo.ecm.core.bulk.message.BulkStatus.State.COMPLETED;
 
 import java.time.Duration;
 import java.util.Collections;
@@ -36,8 +36,8 @@ import javax.ws.rs.core.Response;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.nuxeo.ecm.core.api.DocumentModel;
+import org.nuxeo.ecm.core.bulk.message.BulkStatus;
 import org.nuxeo.ecm.core.bulk.BulkService;
-import org.nuxeo.ecm.core.bulk.BulkStatus;
 import org.nuxeo.ecm.core.bulk.CoreBulkFeature;
 import org.nuxeo.ecm.core.bulk.actions.SetPropertiesAction;
 import org.nuxeo.ecm.core.test.annotations.Granularity;
